Major corrections to teh Autonomous and the creation of new contest op
The autonomous is now including two color sensors on the bottom and use
those for better self alignment. There is still missing data, but that
is to be collected soon and updated soon after. The newest contest Op
allows for the use of our new button pusher and has a funvtion that
allows for some alignment guidence using the color sensors on the bottom